+++ src/plot2D/dialogs/FunctionDialog.cpp
@@ -210,7 +210,7 @@ void FunctionDialog::setCurveToModify(Graph *g, int curve)
         boxFunction->setText(formulas[0]);
         boxFrom->setText(QString::number(c->startRange(), 'g', 15));
         boxTo->setText(QString::number(c->endRange(), 'g', 15));
-        boxPoints->setValue(c->dataSize());
+        boxPoints->setValue(static_cast<int>(c->dataSize()));
     } else if (c->functionType() == FunctionCurve::Polar) {
         optionStack->setCurrentIndex(2);
         boxType->setCurrentIndex(2);
@@ -220,7 +220,7 @@ void FunctionDialog::setCurveToModify(Graph *g, int curve)
         boxPolarParameter->setText(c->variable());
         boxPolarFrom->setText(QString::number(c->startRange(), 'g', 15));
         boxPolarTo->setText(QString::number(c->endRange(), 'g', 15));
-        boxPolarPoints->setValue(c->dataSize());
+        boxPolarPoints->setValue(static_cast<int>(c->dataSize()));
     } else if (c->functionType() == FunctionCurve::Parametric) {
         boxType->setCurrentIndex(1);
         optionStack->setCurrentIndex(1);
@@ -230,7 +230,7 @@ void FunctionDialog::setCurveToModify(Graph *g, int curve)
         boxParameter->setText(c->variable());
         boxParFrom->setText(QString::number(c->startRange(), 'g', 15));
         boxParTo->setText(QString::number(c->endRange(), 'g', 15));
-        boxParPoints->setValue(c->dataSize());
+        boxParPoints->setValue(static_cast<int>(c->dataSize()));
     }
 }
